refactor: remove ElmInlineLink component and update ElmInlineText to support link functionality
