--- packages/core/src/components/inline/ElmInlineLink.stories.tsx
@@ -1,36 +1,0 @@
-import type { Meta, StoryObj } from '@storybook/vue3'
-import ElmInlineLink from './ElmInlineLink.vue'
-
-const meta: Meta<typeof ElmInlineLink> = {
-  title: 'Components/Inline/ElmInlineLink',
-  component: ElmInlineLink,
-  tags: ['autodocs'],
-  args: {}
-}
-
-export default meta
-type Story = StoryObj<typeof meta>
-
-export const Primary: Story = {
-  args: { text: 'elm-inline-link', href: 'https://google.com' }
-}
-
-export const OnClick: Story = {
-  args: { text: 'Open Dialog', onClick: () => alert('clicked!') }
-}
-
-export const OpenInSameTab: Story = {
-  args: {
-    text: 'open in same tab',
-    href: 'https://example.com',
-    openInNewTab: false
-  }
-}
-
-export const Size: Story = {
-  args: {
-    text: 'elm-inline-link',
-    href: 'https://example.com',
-    size: '0.75rem'
-  }
-}
