+++ packages/core/src/components/inline/ElmInlineText.stories.tsx
@@ -26,3 +26,11 @@ export const Background: Story = {
 export const Ruby: Story = {
   args: { text: '国際連合', ruby: 'こくさいれんごう' }
 }
+
+export const Link: Story = {
+  args: {
+    text: 'Google Search',
+    href: 'https://google.com',
+    favicon: 'https://www.google.com/favicon.ico'
+  }
+}
